64bit: changed to using bitwidth insensitive type (hopefully) for bits in union

+++ coast/foundation/base/Anything.h
@@ -515,7 +515,7 @@ protected:
 	union {
 		AnyImpl *fAnyImp;
 		Allocator *fAlloc;
-		unsigned int bits;
+		ptrdiff_t bits;
 	};
 	friend class AnythingTest;
 };
